Trait bmw_evh2::UserContext

pub trait UserContext {
    // Required methods
    fn clone_next_chunk(
        &mut self,
        connection: &mut Connection,
        buf: &mut [u8]
    ) -> Result<usize, Error>;
    fn cur_slab_id(&self) -> usize;
    fn clear_all(&mut self, connection: &mut Connection) -> Result<(), Error>;
    fn clear_through(
        &mut self,
        slab_id: usize,
        connection: &mut Connection
    ) -> Result<(), Error>;
    fn get_user_data(&mut self) -> &mut Option<Box<dyn Any + Send + Sync>>;
    fn set_user_data(&mut self, user_data: Box<dyn Any + Send + Sync>);
}

The crate::UserContext trait is returned on all callbacks specified by the crate::EventHandler. It can be used to read data from the underlying connections, clear slabs that are not needed, and get or set a user_data structure that can be used as a context variable by the caller. The user_data structure is of the type Box<dyn Any + Send + Sync> so the user may use this for practically anything. This value stays consistent accross all callbacks for the thread that is invoked on. Each thread of the crate::EventHandler does store and return a distinct value. See the crate documentation as well for the background information and motivation for this crate as well as examples. See crate::EventHandler for the callback functions.

Required Methods

fn clone_next_chunk( &mut self, connection: &mut Connection, buf: &mut [u8] ) -> Result<usize, Error>

Clone the next chunk of data, if available, that has been read by the crate::EventHandler for this crate::Connection.

Input Parameters

connection - the crate::Connection to get the next chunk from. buf - the mutable slice to clone the data into.

Returns

On success, length in bytes, of data cloned is returned and on failure, [bmw_err::Error] is returned.

Errors

[bmw_err::ErrKind::IllegalArgument] - If the buffer is too small to store the chunk’s data. The maximum size returned is the value configured using [bmw_conf::ConfigOption::EvhReadSlabSize] parameter.

See Also

crate, crate::UserContext, crate::UserContext::clear_all

fn cur_slab_id(&self) -> usize

Retrieve the slab_id of the current slab being processed by this crate::Connection.

Input Parameters

none

Returns

The slab_id of the slab that is currently pointed to by this crate::UserContext. If there are no more slabs associated with this crate::Connection, a number equal to or greater than u32::MAX is returned.

See Also

crate, crate::UserContext, crate::UserContext::clear_through

fn clear_all(&mut self, connection: &mut Connection) -> Result<(), Error>

Clear all slabs that are associated with this crate::Connection.

Input Parameters

connection - The crate::Connection to clear data from.

Returns

On success, unit is returned and on failure, [bmw_err::Error] is returned.

See Also

crate, crate::UserContext, crate::UserContext::clear_through, crate::UserContext::cur_slab_id

fn clear_through( &mut self, slab_id: usize, connection: &mut Connection ) -> Result<(), Error>

Clear all slabs, through the slab_id specified, that are associated with this crate::Connection.

Input Parameters

slab_id - The slab_id of the slab to clear data through. This value can be obtained by calling the crate::UserContext::cur_slab_id as the crate::UserContext::clone_next_chunk function is called. connection - The crate::Connection to clear data from.

Returns

On success, unit is returned and on failure, [bmw_err::Error] is returned.

See Also

crate, crate::UserContext, crate::UserContext::clear_all, crate::UserContext::cur_slab_id

fn get_user_data(&mut self) -> &mut Option<Box<dyn Any + Send + Sync>>

Get the user_data object associated with this thread.

Input Parameters

none

Returns

The user_data which has been specified on a previous call to crate::UserContext::set_user_data.

See Also

crate, crate::UserContext, crate::UserContext::set_user_data

fn set_user_data(&mut self, user_data: Box<dyn Any + Send + Sync>)

Set the user_data object associated with this thread.

Input Parameters

user_data - The value to set for this thread’s context data.

Returns

none

See Also

crate, crate::UserContext, crate::UserContext::get_user_data